+++ src/main/java/com/github/bgomar/bgconsolelogger/tools/ConsoleLoggerSettings.java
@@ -76,7 +76,7 @@ public final class ConsoleLoggerSettings implements PersistentStateComponent<Con
             DEFAULT_PATTERN_18
     ));
 
-    public String version = "0.0.30";
+    public String version = "0.0.31";
 
     public static ConsoleLoggerSettings getInstance() {
         return ApplicationManager.getApplication().getService(ConsoleLoggerSettings.class);
@@ -118,4 +118,3 @@ public final class ConsoleLoggerSettings implements PersistentStateComponent<Con
         }
     }
 }
-